Add authorization in JobOffer controller

+++ MainApp/MainApp/Controllers/JobOfferController.cs
@@ -8,6 +8,7 @@ using MainApp.Models;
 using MainApp.EntityFramework;
 using Microsoft.EntityFrameworkCore;
 using MainApp.Authorization;
+using Microsoft.AspNetCore.Authorization;
 
 namespace MainApp.Controllers
 {
@@ -19,11 +20,12 @@ namespace MainApp.Controllers
         {
             _context = context;
         }
-        
+
         [HttpGet]
+        [Authorize]
         public async Task<IActionResult> Index([FromQuery(Name = "search")] string searchString)
         {
-            Role role = Role.HR;
+            Role role = await AuthorizationTools.GetRoleAsync(User, _context);
             List<JobOffer> searchResult;
             if (string.IsNullOrEmpty(searchString))
             {
@@ -37,27 +39,32 @@ namespace MainApp.Controllers
                     || o.Company.Name.Contains(searchString, StringComparison.OrdinalIgnoreCase))
                     .ToListAsync();
             }
+            string email = AuthorizationTools.GetEmail(User);
             if (role == Role.HR)
             {
                 JobOfferIndexHRView jobOfferIndexHRView = new JobOfferIndexHRView();
                 jobOfferIndexHRView.Offers = searchResult;
-                //jobOfferIndexHRView.HR = thishr;
-                jobOfferIndexHRView.HR = new HR();
+                HR us = _context.HRs.Where(h => h.EmailAddress == email).First();
+                jobOfferIndexHRView.HR = us;
                 return View("IndexHR", jobOfferIndexHRView);
             }
-            else if(role == Role.CANDIDATE)
+            else if (role == Role.CANDIDATE)
             {
                 JobOfferIndexCandidateView jobOfferIndexCandidateView = new JobOfferIndexCandidateView();
                 jobOfferIndexCandidateView.Offers = searchResult;
-                //jobOfferIndexCandidateView.Candidate = thiscandidate;
-                jobOfferIndexCandidateView.Candidate = new Candidate();
+                Candidate us = _context.Candidates.Where(c => c.EmailAddress == email).First();
+                jobOfferIndexCandidateView.Candidate = us;
                 return View("IndexCandidate", jobOfferIndexCandidateView);
             }
             //role == Role.ADMIN
             return View("IndexAdmin", searchResult);
         }
+        [Authorize]
         public async Task<IActionResult> Edit(int? id)
         {
+            Role role = await AuthorizationTools.GetRoleAsync(User, _context);
+            if (role == Role.CANDIDATE)
+                return new UnauthorizedResult();
             if (id == null)
             {
                 return BadRequest($"id shouldn't be null");
@@ -67,6 +74,13 @@ namespace MainApp.Controllers
             {
                 return NotFound($"offer not found in DB");
             }
+            if (role == Role.HR)
+            {
+                string email = AuthorizationTools.GetEmail(User);
+                HR us = _context.HRs.Where(h => h.EmailAddress == email).First();
+                if (us.Id != offer.HR.Id)
+                    return new UnauthorizedResult();
+            }
 
             return View(offer);
         }
@@ -75,12 +89,23 @@ namespace MainApp.Controllers
         [ValidateAntiForgeryToken]
         public async Task<ActionResult> Edit(JobOffer model)
         {
+            Role role = await AuthorizationTools.GetRoleAsync(User, _context);
+            if (role == Role.CANDIDATE)
+                return new UnauthorizedResult();
+
             if (!ModelState.IsValid)
             {
                 return View();
             }
 
             var offer = await _context.JobOffers.FirstOrDefaultAsync(x => x.Id == model.Id);
+            if (role == Role.HR)
+            {
+                string email = AuthorizationTools.GetEmail(User);
+                HR us = _context.HRs.Where(h => h.EmailAddress == email).First();
+                if (us.Id != offer.HR.Id)
+                    return new UnauthorizedResult();
+            }
             offer.JobTitle = model.JobTitle;
             offer.Description = model.Description;
             _context.Update(offer);
@@ -89,13 +114,25 @@ namespace MainApp.Controllers
         }
 
         [HttpPost]
+        [Authorize]
         public async Task<ActionResult> Delete(int? id)
         {
+            Role role = await AuthorizationTools.GetRoleAsync(User, _context);
+            if (role == Role.CANDIDATE)
+                return new UnauthorizedResult();
+
             if (id == null)
             {
                 return BadRequest($"id should not be null");
             }
-
+            var offer = await _context.JobOffers.FirstOrDefaultAsync(x => x.Id == id.Value);
+            if (role == Role.HR)
+            {
+                string email = AuthorizationTools.GetEmail(User);
+                HR us = _context.HRs.Where(h => h.EmailAddress == email).First();
+                if (us.Id != offer.HR.Id)
+                    return new UnauthorizedResult();
+            }
             _context.JobOffers.Remove(new JobOffer() { Id = id.Value });
             await _context.SaveChangesAsync();
             return RedirectToAction("Index");
@@ -115,6 +152,10 @@ namespace MainApp.Controllers
         [ValidateAntiForgeryToken]
         public async Task<ActionResult> Create(JobOfferCreateView model)
         {
+            Role role = await AuthorizationTools.GetRoleAsync(User, _context);
+            if (role == Role.CANDIDATE)
+                return new UnauthorizedResult();
+
             if (!ModelState.IsValid)
             {
                 model.Companies = await _context.Companies.ToListAsync();
@@ -143,21 +184,18 @@ namespace MainApp.Controllers
             var offer = await _context.JobOffers
                 .Include(x => x.Company)
                 .FirstOrDefaultAsync(x => x.Id == id);
-            Role role = Role.HR;
+            Role role = await AuthorizationTools.GetRoleAsync(User, _context);
             if (role == Role.HR)
             {
                 JobOfferDetailsHRView jobOfferDetailsHRView = new JobOfferDetailsHRView();
                 jobOfferDetailsHRView.Offer = offer;
-                //jobOfferIndexHRView.HR = thishr;
-                jobOfferDetailsHRView.HR = new HR()
-                {
-                    Company = offer.Company
-                };
-                //jobOfferDetailsHRView.Applications = await _context.JobApplications.Where(x => x.HR == thishr).ToListAsync();
-                jobOfferDetailsHRView.Applications = await _context.JobApplications.ToListAsync();
+                string email = AuthorizationTools.GetEmail(User);
+                HR us = _context.HRs.Where(h => h.EmailAddress == email).First();
+                jobOfferDetailsHRView.HR = us;
+                jobOfferDetailsHRView.Applications = await _context.JobApplications.Where(ja => ja.JobOffer.HR == us).ToListAsync();
                 return View("DetailsHR", jobOfferDetailsHRView);
             }
-            if(role == Role.ADMIN)
+            if (role == Role.ADMIN)
                 return View("DetailsAdmin", offer);
             return View("DetailsCandidate", offer);
         }
